fix package renaming issue of synapse.api to synapse.rest
