+++ components/esb-tools/plugins/org.wso2.integrationstudio.esb.libraries/build.properties
@@ -144,12 +144,12 @@ bin.includes = META-INF/,\
                lib/swagger-parser-v3-2.0.24.wso2v1.jar,\
                lib/validation-api-1.1.0.Final.jar,\
                lib/jackson-core-2.12.1.jar,\
-               lib/snakeyaml-1.27.jar,\
                lib/swagger-annotations-2.1.6.jar,\
                lib/swagger-models-2.1.6.jar,\
                lib/jackson-annotations-2.12.1.jar,\
                lib/jackson-databind-2.12.1.jar,\
-               lib/jackson-dataformat-yaml-2.12.1.jar,\
                lib/jackson-datatype-jsr310-2.12.1.jar,\
                lib/synapse-core-2.1.7-wso2v142.jar,\
-               lib/smooks-1.5.1.wso2v4.jar
+               lib/smooks-1.5.1.wso2v4.jar,\
+               lib/jackson-dataformat-yaml-2.12.1.jar,\
+               lib/snakeyaml-1.27.jar
